requirements.txt
fastapi
uvicorn
python-dotenv
requests
pytesseract
pdf2image
pillow
python-multipart
cohere
google-generativeai
pinecone-client
